+++ cryptofeed/backends/influxdb.py
@@ -10,8 +10,8 @@ from decimal import Decimal
 import requests
 
 from cryptofeed.backends.backend import (BackendBookCallback, BackendBookDeltaCallback, BackendFundingCallback,
-                                         BackendOpenInterestCallback, BackendTickerCallback, BackendTradeCallback, BackendLiquidationsCallback,
-                                         BackendProfileCallback)
+                                         BackendOpenInterestCallback, BackendTickerCallback, BackendTradeCallback,
+                                         BackendLiquidationsCallback, BackendProfileCallback, BackendTransactionsCallback)
 from cryptofeed.backends.http import HTTPCallback
 from cryptofeed.defines import BID, ASK
 from cryptofeed.exceptions import UnsupportedType
@@ -165,3 +165,8 @@ class LiquidationsInflux(InfluxCallback, BackendLiquidationsCallback):
 
 class ProfileInflux(InfluxCallback, BackendProfileCallback):
     default_key = 'profile'
+
+
+class TransactionsInflux(InfluxCallback, BackendTransactionsCallback):
+    default_key = 'transactions'
+