+++ cryptofeed/backends/rabbitmq.py
@@ -11,7 +11,7 @@ from yapic import json
 
 from cryptofeed.backends.backend import (BackendBookCallback, BackendBookDeltaCallback, BackendFundingCallback,
                                          BackendOpenInterestCallback, BackendTickerCallback, BackendTradeCallback,
-                                         BackendLiquidationsCallback, BackendProfileCallback)
+                                         BackendLiquidationsCallback, BackendProfileCallback, BackendTransactionsCallback)
 
 
 class RabbitCallback:
@@ -103,3 +103,8 @@ class LiquidationsRabbit(RabbitCallback, BackendLiquidationsCallback):
 
 class ProfileRabbit(RabbitCallback, BackendProfileCallback):
     pass
+
+
+class TransactionsRabbit(RabbitCallback, BackendTransactionsCallback):
+    pass
+